Pojo层调整id相关属性，实现点赞实体类

+++ online-study-backend/online-study-pojo/src/main/java/com/study/dto/PostPageQueryDTO.java
@@ -35,7 +35,7 @@ public class PostPageQueryDTO implements Serializable {
     private Integer commentTop;
 
     // 帖子种类Id
-    private Integer categoryId;
+    private Long categoryId;
 
     // 页码数
     private Integer page;
